+++ Sources/CapsomniaCLI/CLICommand.swift
@@ -2,7 +2,7 @@ import Foundation
 import CapsomniaControl
 
 public struct CapsomniaCLICommand {
-    public static let version = "0.1.0"
+    public static let version = "0.1.1"
 
     public init() {}